pyproject.toml
[project]
name = "rce-autocomplete"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "fastapi[standard]>=0.115.8",
    "openai>=1.63.2",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
]

[tool.black]
line-length=79